+++ source/NVDAObjects/window/scintilla.py
@@ -313,7 +313,7 @@ class ScintillaTextInfo(textInfos.offsets.OffsetsTextInfo):
 				tempOffset -= 1
 		return [start, end]
 
-	def collapse(self, end: bool = False):
+	def collapseAndMaybeMoveToNextParagraph(self, end: bool = False):
 		"""Before collapsing to end, if no text is selected, TextInfo is expanded to line.
 		This fixes a bug where next braille line command didn't move the cursor to the last empty line
 		in Notepad++ documents.